Dikoo24

+++ attachments/views.py
@@ -12,8 +12,14 @@ from utility.calendar import PersianCalendar
 from utility.log import leolog
 from django.utils import timezone
 from core.views import CoreContext,PageBriefSerializer,AddRelatedPageForm
-from .repo import LikeRepo,CommentRepo,LinkRepo,DownloadRepo
+from .repo import LikeRepo,CommentRepo,LinkRepo,DownloadRepo, TagRepo
 import json
+
+
+from .repo import ImageRepo,LocationRepo,AreaRepo
+from .serializer import ImageSerializer,LocationSerializer,AreaSerializer,TagSerializer
+       
+
 LAYOUT_PARENT='phoenix/layout.html'
 TEMPLATE_ROOT='attachments/'
 WIDE_LAYOUT="WIDE_LAYOUT"
@@ -45,8 +51,9 @@ def PageCommentsContext(request,page,profile,*args, **kwargs):
     context['comments']=comments  
     context['comments_s']=comments_s  
     if profile is not None:
-        context['add_comment_form']=AddPageCommentForm()
-        context['delete_comment_form']=DeletePageCommentForm()
+        if request.user.has_perm(APP_NAME+'.add_comment'):
+            context['add_comment_form']=AddPageCommentForm()
+            context['delete_comment_form']=DeletePageCommentForm()
     return context
  
 def PageLinksContext(request,page,profile,*args, **kwargs):
@@ -57,7 +64,8 @@ def PageLinksContext(request,page,profile,*args, **kwargs):
     context['links']=links  
     context['links_s']=links_s  
     if profile is not None:
-        context['add_link_form']=AddLinkForm()
+        if profile.user.has_perm(APP_NAME+'.add_link'):
+            context['add_link_form']=AddLinkForm()
     return context
 
 def PageDownloadsContext(request,page,profile,*args, **kwargs):
@@ -68,7 +76,21 @@ def PageDownloadsContext(request,page,profile,*args, **kwargs):
     context['downloads']=downloads  
     context['downloads_s']=downloads_s  
     if profile is not None:
-        context['add_download_form']=AddDownloadForm()
+        if profile.user.has_perm(APP_NAME+'.add_download'):
+            context['add_download_form']=AddDownloadForm()
+    return context
+
+
+def PageTagsContext(request,page,profile,*args, **kwargs):
+    context={}
+    tag_repo = TagRepo(request=request) 
+    tags=tag_repo.list(page_id=page.id)
+    tags_s=json.dumps(TagSerializer(tags,many=True).data)
+    context['tags']=tags  
+    context['tags_s']=tags_s  
+    if profile is not None:
+        if profile.user.has_perm(APP_NAME+'.add_tag'):
+            context['add_tag_form']=AddTagForm()
     return context
 
 
@@ -95,6 +117,23 @@ def PageImagesContext(request,page,profile,*args, **kwargs):
     if profile is not None:
         context['add_image_form']=AddImageForm()
     return context
+
+
+def PageLocationsContext(request,page,*args, **kwargs):
+    context={}
+    location_repo = LocationRepo(request=request) 
+    locations=page.locations.all()
+    locations_s=json.dumps(LocationSerializer(locations,many=True).data)
+    context['locations']=locations  
+    context['locations_s']=locations_s  
+
+    if request.user.has_perm(APP_NAME+'.add_location'):
+        all_locations=location_repo.list()
+        context['all_locations']=all_locations  
+        context['add_page_location_form']=AddPageLocationForm()  
+        context['add_location_form']=AddLocationForm()
+    return context
+
 
 class DownloadView(View):
     def get(self, request, *args, **kwargs): 
@@ -147,6 +186,38 @@ class IndexView(View):
         
       
 
+
+class TagView(View):
+    def get(self, request, *args, **kwargs):
+        me = ProfileRepo(request=request).me
+        tag = TagRepo(request=request).tag(*args, **kwargs)
+        if tag is None:
+            raise Http404
+        context=getContext(request=request)
+        context['tag']=tag
+        page_tags=tag.pages.all()
+        context['page_tags']=page_tags
+        pages=tag.pages.all()
+        pages_s=json.dumps(PageBriefSerializer(pages,many=True).data)
+        context['pages']=pages
+        context['pages_s']=pages_s
+        context['expand_pages']=True
+
+        return render(request,TEMPLATE_ROOT+"tag.html",context)
+       
+       
+
+class TagsView(View):
+    def get(self, request, *args, **kwargs):
+        me = ProfileRepo(request=request).me
+        tags = TagRepo(request=request).list(*args, **kwargs)
+        tags_s=json.dumps(TagSerializer(tags,many=True).data)
+        context=getContext(request=request)
+        context['tags']=tags
+        context['tags_s']=tags_s
+
+        return render(request,TEMPLATE_ROOT+"tags.html",context)
+       
       
       
 class LinksView(View):
@@ -186,9 +257,6 @@ class CommentsView(View):
         
       
 
-from .repo import ImageRepo
-from .serializer import ImageSerializer
-       
 class ImagesView(View):
     def get(self, request, *args, **kwargs): 
         context=getContext(request=request)
@@ -220,3 +288,61 @@ class ImageDownloadView(View):
         if image is None:
             raise Http404
         return image.download_response()
+
+
+
+class LocationView(View):
+    def get(self, request, *args, **kwargs):
+        context = getContext(request)
+        location = LocationRepo(request=request).location(*args, **kwargs)
+        context['location'] = location
+        context['location_s'] = json.dumps(LocationSerializer(location).data)
+
+        return render(request, TEMPLATE_ROOT+"location.html", context)
+  
+
+class AreaView(View):
+    def get(self, request, *args, **kwargs):
+        context = getContext(request)
+        area = AreaRepo(request=request).area(*args, **kwargs)
+        context['area'] = area
+        for installed_app in context['installed_apps']:
+            if installed_app['name']=='market':
+                from market.repo import SupplierRepo,CustomerRepo
+                from market.serializers import SupplierSerializer,CustomerSerializer
+              
+                suppliers=SupplierRepo(request=request).list(region_id=area.id)
+                context['suppliers'] = suppliers
+                suppliers_s=json.dumps(SupplierSerializer(suppliers,many=True).data)
+                context['suppliers_s']=suppliers_s
+
+                
+                customers=CustomerRepo(request=request).list(region_id=area.id)
+                context['customers'] = customers
+                customers_s=json.dumps(CustomerSerializer(customers,many=True).data)
+                context['customers_s']=customers_s
+        context['area_s'] = json.dumps(AreaSerializer(area).data)
+        return render(request, TEMPLATE_ROOT+"area.html", context)
+
+
+class AreasView(View):
+    def get(self, request, *args, **kwargs):
+        context = getContext(request)
+        areas = AreaRepo(request=request).list(*args, **kwargs)
+        context['areas'] = areas
+        context['areas_s'] = json.dumps(AreaSerializer(areas,many=True).data)
+        if request.user.has_perm(APP_NAME+".add_area"):
+            context['add_area_form']=AddAreaForm()
+            context['colors']=(color[0] for color in ColorEnum.choices)
+        return render(request, TEMPLATE_ROOT+"areas.html", context)
+
+
+class LocationsView(View):  
+    def get(self, request, *args, **kwargs):
+        context = getContext(request)
+        locations = LocationRepo(request=request).list(*args, **kwargs)
+        context['locations'] = locations
+        context['locations_s'] = json.dumps(LocationSerializer(locations,many=True).data)
+        if request.user.has_perm(APP_NAME+".add_location"):
+            context['add_location_form']=AddLocationForm()
+        return render(request, TEMPLATE_ROOT+"locations.html", context)
